# [sgen] Optimize LOS for better locality and parallelization.