--- a/mono/sgen/sgen-array-list.h
+++ b/mono/sgen/sgen-array-list.h
@@ -49,6 +49,38 @@ typedef struct {
 	int mem_type; /* sgen internal mem type or -1 for malloc allocation */
 } SgenArrayList;
 
+#if defined(__GNUC__)
+static inline guint32
+sgen_clz (guint32 x)
+{
+	return __builtin_clz (x);
+}
+#elif !defined(ENABLE_MSVC_LZCNT) && defined(_MSC_VER)
+static inline guint32
+sgen_clz (guint32 x)
+{
+	gulong leading_zero_bits;
+	return _BitScanReverse (&leading_zero_bits, (gulong)x) ? 31 - leading_zero_bits : 32;
+}
+#elif defined(ENABLE_MSVC_LZCNT) && defined(_MSC_VER)
+static inline guint32
+sgen_clz (guint32 x)
+{
+	return __lzcnt (x);
+}
+#else
+static inline guint32
+sgen_clz (guint32 x)
+{
+	guint count = 0;
+	while (x) {
+		++count;
+		x >>= 1;
+	}
+	return 32 - count;
+}
+#endif
+
 /*
  * Computes floor(log2(index + MIN_BUCKET_SIZE)) - 1, giving the index
  * of the bucket containing a slot.
@@ -56,17 +88,7 @@ typedef struct {
 static inline guint32
 sgen_array_list_index_bucket (guint32 index)
 {
-#ifdef __GNUC__
-	return CHAR_BIT * sizeof (index) - __builtin_clz (index + SGEN_ARRAY_LIST_MIN_BUCKET_SIZE) - 1 - SGEN_ARRAY_LIST_MIN_BUCKET_BITS;
-#else
-	guint count = 0;
-	index += SGEN_ARRAY_LIST_MIN_BUCKET_SIZE;
-	while (index) {
-		++count;
-		index >>= 1;
-	}
-	return count - 1 - SGEN_ARRAY_LIST_MIN_BUCKET_BITS;
-#endif
+	return CHAR_BIT * sizeof (index) - sgen_clz (index + SGEN_ARRAY_LIST_MIN_BUCKET_SIZE) - 1 - SGEN_ARRAY_LIST_MIN_BUCKET_BITS;
 }
 
 static inline guint32