Running parallel scan in minor GC doesn't scale out well over multiple cores when the LOS list is getting big. One of the reasons why this happens is the way each job travers the LOS list. Currently, all jobs scan the complete LOS list, skipping jobs not matching their individual job index.

Problem with this is that each thread needs to touch every element in the list, and by default, we  create number of cores * 4 jobs scanning LOS objects. That means we will touch every element 32 times on an 8 core machine, moving a lot of memory between cache levels, increasing the total scan time for the complete LOS scan job.

This is especially problematic on hardware with slower memory bus and smaller caches, but with many cores. On such hardware the accumulated scan time can be bigger when parallelize compared to just having one thread doing the scan, but all depends on how many objects each job needs to scan. In one scenario we had a major heap of 3 GB and ~600 MB of LOS , split on ~ 12500 LOS objects. Doing minor GC on that heap took ~20 ms running without parallelization. With default parallelization we were down to ~19 ms and the time scanning the LOS objects didn't improve. NOTE, the more work each job gets (more memory to scan) we will get a better default parallelization, but still a lot of room for improvement.

Changing from a normal linked list to a SgenArrayList makes it more effective to parallelize the scan jobs and since the list is using sequential memory, this will also reduce cache misses for each thread iterating LOS objects. Tagging objects in SgenArrayList with reference information reduce cache misses dramatically in cases where objects not including references exists on the list, since scanning jobs can skip these items without any need to touch object memory causing cache misses.

Same scenario running with this fix gives us minor GC pause times, without parallelization in ~ 7ms (down from ~ 20ms) and with parallelization, < 5ms (down from 19ms). NOTE, the more work each job gets (more memory to scan) the better result using parallelization, so the delta will increase between using and not using parallelization, this scenario primarily stress the extreme case where most LOS objects doesn't have any references.

In total, this optimization reduce minor GC pause time from ~ 19ms when using parallelization down to < 5ms, that's close to 4x improvement in minor GC pause times.

This is the first PR in a series of optimizations around minor GC parallelization that have made a significant performance increase, both on local tests as well as in real world large applications very sensitive to minor GC pause times, like games. When applying all upcoming PR's on top of the same scenario, we are able to complete minor GC in ~ 2.5 ms using parallel minor GC, that initially took ~ 20 ms, that's 8x improvements in minor GC pause times.