@BrzVlad 2019-10-05
Overall I have mixed feelings about this. I don't like that much that it complicates a little bit the large object list layout. It is also hard to tell how this will behave for long running programs, in theory this might need some load balancing sometimes. Also the table size is a little bit arbitrary and it would be nice if we didn't have to deal with that as a compile time constant.

I'm wondering if it is worth trying out a simpler approach. We could keep track of the number of los objects at all time (alloc++/free--). When submitting the LOS scan jobs we could instead submit a job whose job is to produce all the smaller jobs. It could pass to each job the LOSObject start and the number of objects to scan. As it traverses a section of the list, it would already enqueue a new job, so once, a small section of the list has been traversed, parallelization already starts. With this approach we keep the simple los layout, we always do balancing of the jobs and the overhead of 2 list traversals instead of one (with one also being parallelized) should be negligible (way better than 32 anyway).

The parallelization model for the los scanning is already not great, because we don't take into account the size of each object. For example scanning a 8KB object is very different than scanning an 8MB object, but we consider them all the same. Going with the approach in this PR might make it also awkward to implement that in the future. While if we have the simple job enqueuing approach, we could easily expand it to divide jobs based on total size to be scanned, instead of object count.

@lateralusX I think it is worth trying out this approach and see if it has the same benefits as the one in this PR.