@jaykrell 2019-10-07
Fair enough. Comment is more for anyone implementing a hash table: 
 - hash value should be stored with full value in the table
 - hash functions should return 64 bits

So that values in the same bucket ("collision"), can be compared for equality more efficiently, by comparing their full hash value before resorting to comparing their full value.