@lateralusX 2019-10-07
@BrzVlad Couple of thoughts:

Size of list could be dynamically allocated based on the job split (or configuration), if we think that the default bucket size is arbitrary (use simliar default size for other implementations, SGEN_ARRAY_LIST_BUCKETS).

I ran with this for sometime with random allocations of LOS objects and the objects seems to distribute quite well over the buckets, but of course, in theory some of the list might get more objects than other, but even if we could get bad distribution over time, it will still be split out over more list compared to what we have today. Having a larger list count compared to scan jobs would also assist that scenario a little, reducing the risk of unbalanced work loads/thread. Do we have any tests that we could use to investigate if this is something more than a potential theoretical issue?

I also thought around better split of LOS scan jobs taken into account size of objects in the distribution and if we are going to solve that, then we will need a different split, but in order to get that working we will need to know total size of LOS objects (something we probably already know), and we will need to look through the card table for each object to find out how much of that object that really needs to be scanned. Is that something we do when parallelize the major heap scan today, or could individual threads take a higher hit if they ends up with blocks that have more memory that needs to be scanned (based on marked cards)?

Regarding parallelization starting early, one thing I notice when working on this, is that none of the jobs queued in enqueue_scan_remembered_set_jobs will be picked up by thread pool until sgen_workers_start_all_workers, so currently none of the jobs can be started until we call that method since it sets the context->started to true and none of the workers will pickup jobs until that gets set.

I could benchmark alternative ways to parallelize this (like the suggestion above), but I suspect that the first traversal of the whole list (based on below observations), will take more time than expected. Did a test on targeted hardware around this. Just iterating the LOS list and calculate number of items + total size using one thread takes ~2.5 - 2.8 ms:

`Scanning LOS list calculating size: items 11102, 604015144 bytes, 2540 usecs`

The complete minor GC completes in ~6 ms with all optimizations in place. Paying 2.5 ms to use a different algorithm is a rather high price to pay. Also, that price will probably be payed more than once when we try to parallelize the LOS list in more places.

I have another PR on the way that builds on this way of splitting the LOS list into buckets, that parallelize the copy/clear of card table, that PR reduce that part from 8 ms down to 3 ms in a benchmark with 3GB major and 600 MB LOS and if we do a different algorithm, then that parallelization will be harder to accomplish since its mainly about iterating the list, comparing all found objects against card table to copy/clear marked cards, nothing will be scanned, so first iterating through the list will cost 2.5 ms in copy step and (unless we preserve the split) an additional 2.5 ms in clear, defeating pretty much what we gain in parallelize this phase in the first place.

So to summarize my thoughts:

Yes this complicates the layout of LOS list a little, but most of the code is still the same, we could even restore the old behavior by setting a size to 1 and the compiler will most likely reduce the outer loop over buckets, just leaving the inner loop, giving the same behavior as we had before. I even preserved the old way of splitting up jobs if we only have a list of size 1. If the default hard list size is an issue, we could switch into a configurable size that more easily could be tweaked.

This way of splitting the LOS list also enables for further parallelization of other tasks in minor GC, like copy/clear (mentioned above). If we change this and do the split in a different way, like running through the list once to split up the jobs dynamically as part of minor GC, then we might need to implement logic to preserve the split or we will take the hit multiple times during minor GC. If we weight in scan size as well, then the split might be different depending on what part of minor GC that we parallelize, so we might end up paying the prices more than once, and that will be a significant cost to pay and it will complicate the implementation.

The price of iterating the LOS list could be quite high, depending on hardware and number of items in the list. In the scenario I have been looking at, it's 2.5 - 2.8 ms and we are currently (with all optimizations) at minor GC pause times around ~6 ms (down from 30 ms in my local benchmark). If I stressed this a little, just allocating 8001 bytes up to a total of 600 MB LOS, then iterating through the list of 75000 LOS objects will take 14 ms, but the distribution looked OK despite the same allocation size, all lists had at least 2000 objects, most list in the range 2200 - 2500. Increasing the 

Could an alternative be to switch to configurable list size, and default to 1 (to preserve current behavior) and then have an option to increase the size meaning that we can experiment more easily with this optimization + opening up for the other PR's building on this? I also have some additional logging in later PR's that can output the split over the list on higher log levels making it possible to analyze the split over time.